TST: create pandas/tests/arrays to consolidate pandas-extension array testing (#21795)
