+++ pandas/tests/arrays/categorical/test_algos.py
