+++ pandas/tests/arrays/categorical/test_analytics.py
