+++ pandas/tests/arrays/categorical/test_dtypes.py
