+++ pandas/tests/arrays/categorical/test_missing.py
