+++ pandas/tests/arrays/categorical/test_sorting.py
