+++ pandas/tests/arrays/categorical/test_warnings.py
